Hard-code 'no' domain to refer to 'nb'.
Once and for all, let's hardcode the Norwegian `no` domain to
refer to Norwegian Bokmål `nb` for localization purposes.
In the most ideal case, this kind of case could be handled by
the Accept-Language header, but the PCS service currently uses
this header only for languages that have variants, and `nb` is
not a variant, and improving our usage of Accept-Language would
be part of a much larger refactor.

Bug: T384572
Change-Id: I5a1b474b930ae50655813b49d54fea0995518d54