Use ResourceLoader instead of local copies of styles.
* Remove all local copies of CSS styles from other extensions, skins,
etc., and replace them with a call to ResourceLoader. The build step
for generating the CSS bundle (base.css) is the same as before: npm
run build:css.
* The RL call is made by a new script, which will be executed
manually whenever we want to update the styles to the latest from
upstream. Eventually we can consider an automated method of running
this script periodically (weekly?), combined with tests.
* The script has a basic sanity check for problematic modules, in
case the name or structure of any of the modules we're requesting
changes in the future.
* All of the local logic and styles for attempting to manipulate the
DOM and CSS to accommodate specific dark mode issues has been
removed, and is now delegated to upstream styles.
* With any luck, this patch will allow us to close most, if not all,
tasks related to dark mode presentation, unifying our dark mode with
mobile web, and allowing us to direct editors of articles to amend
their content to display correctly in dark mode on both web and apps.

Bug: T369435
Bug: T252904
Bug: T367485
Change-Id: I4f436add909067bedf673fbe9e42254060648c20