Update styles from Cite extension.
This repo contains local copies of styles from various extensions,
so it's no surprise that some of them are quite out-of-date.

This patch brings the styles from the Cite extension to the current
version, but since this repo still hasn't been updated to use Codex
(being done/discussed in another patch) we need to "backport" some
of the elements to not use Codex-specific stuff.

Bug: T393134
Change-Id: Ia57b7d1df774cab5a364bf5183cda5ec2dda0eb0