Fix clicking of references and reference groups.
Even though the previous patch was CSS-only, it apparently impacts
the way that click events are passed down to the javascript code,
and the element that actually receives the click event has changed.

What was then the parent is now the grandparent.

Also fixes tests that expect the old html structure in the fixture.

Bug: T393134
Change-Id: I60e97a5b952dac3ec2ea03c6948fd1cfd4cb822e