Fix counting of lower-alpha notes.
The output from Cite has further diverged from the legacy logic that
PCS expects the output to be. Therefore we need a few additional
overrides for now. Eventually we should refactor our logic that
restructures references, and simplify it to make as few modifications to
the original output as possible.

Bug: T396741
Change-Id: Ide7d468beb526fce22a8377c34db27ed6b85d976