Rollout updated javascript only.
This holds the built CSS files at their previous versions, so that only
the updated PCS javascript is rolled out for now.
This will allow clients to uptake the new pcs javascript, over the
course of the cache lifetime of these files (24 hours), which will
ensure that when we roll out the CSS, the javascript will definitely
be updated. (Some client versions have quirks regarding fetching
resources from the network, and will explicitly read from local cache
up to the max-age of the resource.)

Change-Id: I7f940352964747d780f5619bdd91dce96362975c