When widening images, take minimum width and height into account.
Certain images are being widened that shouldn't be.
For example, in the article [[Jannik Sinner]], under "Awards and
honours", the "orders" image becomes incorrectly widened.

We already had logic that widens images only if they are larger
than a minimum width, but let's also make sure they are taller
than a minimum height.

Change-Id: I28b31709b269781dc86bfe6d23dd9e1bf230cc4f