Support subreferences (squashed)
For background about subreferencing:
https://meta.wikimedia.org/wiki/WMDE_Technical_Wishes/Sub-referencing

Example page using subreferencing:
https://test.wikipedia.org/wiki/CiteDetails

This patch implements full support for subreferences, which have
footnote markers like "[1.1]" and appear as a nested reference list
under their associated main reference.

Test fixture data in this patch is taken from revision:
https://test.wikipedia.org/w/index.php?title=CiteDetails&oldid=673412

This patch only impacts normal references by one chage, that reflist
items are nested with one more level of <div> element.  This is so
that subreferences (when present) can appear as a sibling of the main
footnote body.

Clicking on a subreference footnote marker sends the usual "reference"
event but its html payload consists of the main ref concatenated with
the subref body.

There is a soft dependency on Ib0492135a568e, which will use data in
the reflist to render ref item numbers.  Without that patch to
parsoid, subreferences will still display but some main refs (if not
used in the document) will have an up arrow as their backlink marker.

Squashes the following patches:
    eb89fa Show subrefs in reference list
    7358f3 Add subreference tests for pagelib ReferenceCollection
    7a9c6d [refactor] Split collectRefText function
    16700c Concatenate main and subref for preview
    59bd73 Reuse test fixture from pagelib
    e4a595 Use data-mw-footnote-number if available
    e150ee Subrefs begin on a new row

Bug: T395877
Change-Id: I7d90079516f72b6ed27e3f37bbbe60dc5d8ceaaa